health_check_using_rpc: put health check related code into separate file
